+++ regression/reference/clibrary/wrapClibrary.c
@@ -57,6 +57,100 @@ void CLI_function4a_bufferify(const char * arg1, int Larg1,
 // splicer end function.function4a_bufferify
 }
 
+// void bindC2(const char * name +intent(in)+len_trim(Lname))
+/**
+ * \brief Rename Fortran name for interface only function
+ *
+ * This creates a Fortran bufferify function and an interface.
+ */
+void CLI_bind_c2_bufferify(const char * name, int Lname)
+{
+// splicer begin function.bind_c2_bufferify
+    char * SH_name = (char *) malloc(Lname + 1);
+    memcpy(SH_name, name, Lname);
+    SH_name[Lname] = '\0';
+    bindC2(SH_name);
+    free(SH_name);
+    return;
+// splicer end function.bind_c2_bufferify
+}
+
+// int passAssumedTypeBuf(void * arg +assumedtype+intent(in), const char * name +intent(in)+len_trim(Lname))
+/**
+ * \brief Test assumed-type
+ *
+ * A bufferify function is created.
+ * Should only be call with an C_INT argument, and will
+ * return the value passed in.
+ */
+int CLI_pass_assumed_type_buf_bufferify(void * arg, const char * name,
+    int Lname)
+{
+// splicer begin function.pass_assumed_type_buf_bufferify
+    char * SH_name = (char *) malloc(Lname + 1);
+    memcpy(SH_name, name, Lname);
+    SH_name[Lname] = '\0';
+    int SHC_rv = passAssumedTypeBuf(arg, SH_name);
+    free(SH_name);
+    return SHC_rv;
+// splicer end function.pass_assumed_type_buf_bufferify
+}
+
+// void callback3(const char * type +intent(in)+len_trim(Ltype), void * in +assumedtype+intent(in), void ( * incr)(int *) +external+intent(in)+value)
+/**
+ * \brief Test function pointer
+ *
+ * A bufferify function will be created.
+ */
+void CLI_callback3_bufferify(const char * type, int Ltype, void * in,
+    void ( * incr)(int *))
+{
+// splicer begin function.callback3_bufferify
+    char * SH_type = (char *) malloc(Ltype + 1);
+    memcpy(SH_type, type, Ltype);
+    SH_type[Ltype] = '\0';
+    callback3(SH_type, in, incr);
+    free(SH_type);
+    return;
+// splicer end function.callback3_bufferify
+}
+
+// int passStruct2(Cstruct1 * s1 +intent(in), const char * name +intent(in)+len_trim(Lname))
+/**
+ * Pass name argument which will build a bufferify function.
+ */
+int CLI_pass_struct2_bufferify(Cstruct1 * s1, const char * name,
+    int Lname)
+{
+// splicer begin function.pass_struct2_bufferify
+    char * SH_name = (char *) malloc(Lname + 1);
+    memcpy(SH_name, name, Lname);
+    SH_name[Lname] = '\0';
+    int SHC_rv = passStruct2(s1, SH_name);
+    free(SH_name);
+    return SHC_rv;
+// splicer end function.pass_struct2_bufferify
+}
+
+// Cstruct1 * returnStructPtr2(int ifield +intent(in)+value, const char * name +intent(in)+len_trim(Lname))
+/**
+ * \brief Return a pointer to a struct
+ *
+ * Generates a bufferify C wrapper function.
+ */
+Cstruct1 * CLI_return_struct_ptr2_bufferify(int ifield,
+    const char * name, int Lname)
+{
+// splicer begin function.return_struct_ptr2_bufferify
+    char * SH_name = (char *) malloc(Lname + 1);
+    memcpy(SH_name, name, Lname);
+    SH_name[Lname] = '\0';
+    Cstruct1 * SHC_rv = returnStructPtr2(ifield, SH_name);
+    free(SH_name);
+    return SHC_rv;
+// splicer end function.return_struct_ptr2_bufferify
+}
+
 // Release C++ allocated memory.
 void CLI_SHROUD_memory_destructor(CLI_SHROUD_capsule_data *cap)
 {