+++ regression/reference/clibrary/wrapClibrary.h
@@ -23,6 +23,7 @@
 #ifndef WRAPCLIBRARY_H
 #define WRAPCLIBRARY_H
 
+#include "clibrary.h"
 #include "typesClibrary.h"
 
 // splicer begin C_declarations
@@ -30,5 +31,19 @@
 
 void CLI_function4a_bufferify(const char * arg1, int Larg1,
     const char * arg2, int Larg2, char * SHF_rv, int NSHF_rv);
+
+void CLI_bind_c2_bufferify(const char * name, int Lname);
+
+int CLI_pass_assumed_type_buf_bufferify(void * arg, const char * name,
+    int Lname);
+
+void CLI_callback3_bufferify(const char * type, int Ltype, void * in,
+    void ( * incr)(int *));
+
+int CLI_pass_struct2_bufferify(Cstruct1 * s1, const char * name,
+    int Lname);
+
+Cstruct1 * CLI_return_struct_ptr2_bufferify(int ifield,
+    const char * name, int Lname);
 
 #endif  // WRAPCLIBRARY_H