+++ shroud/wrapp.py
@@ -2444,7 +2444,9 @@ class ToImplied(todict.PrintNode):
 
     expression has already been checked for errors by generate.check_implied.
     Convert functions:
-      size  -  PyArray_SIZE
+      size     -  PyArray_SIZE
+      len      -  strlen
+      len_trim -
     """
 
     def __init__(self, expr, func):
@@ -2456,12 +2458,29 @@ class ToImplied(todict.PrintNode):
         # Look for functions
         if node.args is None:
             return node.name
+        ### functions
         elif node.name == "size":
             # size(arg)
             argname = node.args[0].name
             #            arg = self.func.ast.find_arg_by_name(argname)
             fmt = self.func._fmtargs[argname]["fmtpy"]
             return wformat("PyArray_SIZE({py_var})", fmt)
+        elif node.name == "len":
+            # len(arg)
+            argname = node.args[0].name
+            #            arg = self.func.ast.find_arg_by_name(argname)
+            fmt = self.func._fmtargs[argname]["fmtpy"]
+            # XXX - need code for len_trim
+            return wformat("strlen({cxx_var})", fmt)
+        elif node.name == "len_trim":
+            # len_trim(arg)
+            argname = node.args[0].name
+            #            arg = self.func.ast.find_arg_by_name(argname)
+            fmt = self.func._fmtargs[argname]["fmtpy"]
+            return wformat("strlen({cxx_var})", fmt)
+            # XXX - need code for len_trim
+            return wformat("ShroudLenTrim({cxx_var}, strlen({cxx_var}))", fmt)
+            #c_helper="ShroudLenTrim"
         else:
             return self.param_list(node)