+++ CHANGELOG.md
@@ -8,9 +8,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added format dictionary field template_suffix.
+- Added the **F_create_bufferify_function** option.
+- Support `true` and `false` with implied attribute.
+- Process 'void **' argument as 'type(C_PTR)' by reference.
+- Added assumedtype attribute to use Fortran assumed-type declaration,
+  `type(*)`, for `void *` arguments.
+- Added external attribute to use with function pointers.
+  Uses EXTERNAL statement instead of abstract interface to allow any function
+  to be used as the dummy argument since the interface is not preserved.
 
 ### Fixed
 - C++ function arguments which pass a class by value.
+- C struct arguments passed to a bufferify function.
+  It was using C++ casting.  Instead there is no C wrapper for a struct
+  like there is with a Class since it is already compatible with Fortran.
 
 ## v0.10.1 - 2018-08-07
 ### Fixed